+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/AdminApi/GameFinderController.cs
@@ -1,0 +1,59 @@
+﻿// MIT License
+//
+// Copyright (c) 2019 Stormancer
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+using Microsoft.AspNetCore.Mvc;
+using Stormancer.Core;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Stormancer.Server.Plugins.GameFinder.AdminApi
+{
+
+    [ApiController]
+    [Route("_gamefinder")]
+    internal class GameFinderController : ControllerBase
+    {
+        private readonly ISceneHost _scene;
+
+        public GameFinderController(ISceneHost scene)
+        {
+            _scene = scene;
+        }
+
+        public async Task<IActionResult> GetStatus(string gamefinderId, CancellationToken cancellationToken)
+        {
+            await using var scope = _scene.CreateRequestScope();
+
+            var gameFinder = scope.Resolve<GameFinderProxy>();
+
+            var metrics = await gameFinder.GetStatus(gamefinderId, cancellationToken);
+
+            return Ok(metrics);
+
+
+        }
+    }
+}